[Chore 159775689] Add test to test post question
