+++ run.py
@@ -1,9 +1,10 @@
 import os
-from api.resources.questions import create_app
 
-config_name = os.getenv('APP_SETTINGS') # config_name = "development"
+from api.resources.app import app
 
-app = create_app(config_name)
+# config_name = os.getenv('APP_SETTINGS') # config_name = "development"
+
+# app = create_app(config_name)
 
 if __name__ == '__main__':
     app.run(debug=True)
